CP-12512 A switch to set stunnel protocol version
Before this change, when we wrote client-mode configuration for stunnel
we did not specify SSL/TLS version or ciphersuites, so it was using
defaults (e.g. SSLv3 for stunnel 4.15, TLSv1.0 for stunnel 4.56)

Now we specify these things: TLSv1.2 protocol and TLSv1.2 ciphersuites
are specified, or if the legacy flag is set, we allow all TLS protocol
versions (but not SSL) and we add the ciphersuites that were accepted
for incoming connections in older XenServer versions (6.5 and earlier).

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>
(cherry picked from commit acea71f37ea964483c44c8853277343a7043e93b)

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>